test(lifted-stark): replace bus-mock test with minimal eval_external test
The bus-flavored test_bus.rs mocked multiset + logup identities only to
exercise MultiAir::eval_external. Drop the bus framing entirely: a single
aux column, one aux value, one external assertion. The crate has no bus
abstraction, so the test no longer pretends to.

Keeps the cross-layer error test (eval_external ReductionError surfacing
as VerifierError::Reduction); renamed to test_external_assertions.rs.
